pyproject.toml
[project]
name = "edu-assistant"
version = "0.1.0"
description = "Simple project for educational AI-assistant"
requires-python = ">=3.12,<3.13"
dependencies = [
    "fastapi[standard,standart]==0.135.3",
    "loguru==0.7.3",
    "openai==2.30.0",
    "pydantic==2.12.5",
    "python-dotenv==1.2.2",
    "python-multipart>=0.0.26",
    "pyyaml==6.0.3",
    "uvicorn[standard,standart]==0.44.0",
]

[build-system]
requires = ["uv_build>=0.11.3,<0.12"]
build-backend = "uv_build"

